requirements.txt
asgiref==3.3.1
cassandra-driver==3.24.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.4
django-cassandra-common==0.0.2
django-cassandra-engine==1.6.1
django-filter==2.4.0
djangorestframework==3.12.2
drf-yasg==1.20.0
geomet==0.2.1.post1
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==20.8
pyparsing==2.4.7
pytz==2020.4
requests==2.25.1
ruamel.yaml==0.16.12
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
